finzalizacion de api(funcional)

+++ Api_Instrumentos/Models/Instrumento.cs
@@ -14,8 +14,8 @@ namespace Api_Instrumentos.Models
         public string Categoria { get; set; }
         public string Marca { get; set; }
         public string Provedor { get; set; }
-        public double precio_compra { get; set; }
-        public double precio_venta { get; set; }
+        public decimal precio_compra { get; set; }
+        public decimal precio_venta { get; set; }
         public int stock { get; set; }
         public int stock_minimo { get; set; }
         public string descripcion { get; set; }
